==> 1_data_processing/02_BROAD_ID_mapping.ipynb <==
## Unified CCLs model ID as BROAD ID

In this step, **model ID (either Sanger Model ID or BROAD ID) would be unified as BROAD ID**. 

**Input**
- **Cell line model annotation (e.g., Sanger model ID, BROAD ID etc.)**: model_list_20230307.csv (Version: 23Q2; https://depmap.org/portal/download/all/)
- Transient Sanger CNV data in previous step: **cnv_sanger_entrezID.csv**
- Transient CRISPR gene effect data in previous step: **crispr_broad_entrezID.csv**

**Output**
- Sanger CNV data with mapped BROAD ID: **cnv_sanger_entrezID_broadID.csv**
- CRISPR gene effect data with mapped BROAD ID: **crispr_broad_entrezID_broadID.csv**

## Import modules
import numpy as np
import pandas as pd

## Import model information for downstream data integration
model = pd.read_csv('/Users/amy/Desktop/SyntheticLethalityProject/sources/model_list_20230307.csv', index_col = None)

## Subset the data for useful columns
model_col = model[['model_id', 'BROAD_ID']]
model_col = model_col.rename(columns = {'model_id': 'SangerModelID', 'BROAD_ID': 'ModelID'})
model_col = model_col.dropna()

## Previous Sanger CNV data and CRISPR gene effect data
cnv_sanger = pd.read_csv('/Users/amy/Desktop/SyntheticLethalityProject/1_data_processing/01_entrez_ID_mapping/cnv_sanger_entrezID.csv', index_col = None, low_memory=False)
crispr_broad = pd.read_csv('/Users/amy/Desktop/SyntheticLethalityProject/1_data_processing/01_entrez_ID_mapping/crispr_broad_entrezID.csv', index_col = None)

## Merge BROAD ID to each data
crispr_broad = pd.merge(model_col, crispr_broad, on = ['ModelID'], how = 'right')
cnv_sanger = pd.merge(model_col, cnv_sanger, on = ['SangerModelID'], how = 'right')
## NA may be presented in Model ID. Hence, we should drop these NA value out. 
cnv_sanger = cnv_sanger.dropna(subset = ['ModelID'])
crispr_broad = crispr_broad.dropna(subset = ['SangerModelID'])

## Rename the column 
cnv_sanger = cnv_sanger.rename(columns={'ModelID':'BROAD_ID'})
crispr_broad = crispr_broad.rename(columns={'ModelID':'BROAD_ID'})


## After mapping to the model identifier list
## The overview of CNV data
print("Number of cell lines in CNV data:", cnv_sanger.shape[0])
print("Number of unique cell lines from Sanger source:", cnv_sanger[cnv_sanger.source == 'Sanger'].BROAD_ID.unique().shape[0])
print("Number of unique cell lines from Broad source:", cnv_sanger[cnv_sanger.source == 'Broad'].BROAD_ID.unique().shape[0])

## The overview of CRISPR data 
print("Number of cell lines in CRISPR data:", crispr_broad.shape[0])
print("Number of unique cell lines in CRISPR data:", crispr_broad.BROAD_ID.unique().shape[0])

cnv_sanger[:2]

crispr_broad[:2]

Save the file with mapped with BROAD ID for further processing.

## Sanger CNV
cnv_sanger.to_csv('/Users/amy/Desktop/SyntheticLethalityProject/1_data_processing/02_BROAD_ID_mapping/cnv_sanger_entrezID_broadID.csv', index=False)
## BROAD CRISPR gene effect
crispr_broad.to_csv('/Users/amy/Desktop/SyntheticLethalityProject/1_data_processing/02_BROAD_ID_mapping/crispr_broad_entrezID_broadID.csv', index=False)
